isolate `Location`
This commit introduces `Location`, a replacement for the old
`Position`, in a new dedicated file. This type is intended to be used
across different representations, and therefore contains the filename,
despite it being a bit unnecessarily verbose when it comes to
tokens. But, that way, we don't have to translate a token's location
into a symbol's location down the line: it will be the same type.

`WithLocation` is just a temporary type alias, that will likely be
changed or removed. If we end up deciding on a higher-kinded AST, it
will be made into a proper newtype instead. Otherwise, we might
prefer avoiding the type alias and use tuples / constructor values.

This commit does NOT compile, as the stream was interrupted before we
could even try to fix compilation issues.