fix lexer to handle comments and identifiers
Due to how we use a "default" rule to parse identifiers if all else
fails, an identifier that starts with the same string as a keyword was
parsed as the latter. The fix for this is to only successfully parse a
keyword if it's followed by a boundary: a symbol or whitespace.