fix last ambiguity in language
This is ugly and annoying: we are not able to distinguish between an
expression that ends with an identifier and is followed by the opening
brace of a block, and a struct expression which is the type name (an
identifier) followed by an opening brace. As a result, we can't
distinguish between the two, and something like `if foo { bar }` would
be parsed as creating a struct foo in the condition of the `if`.

This could be fixed by rewriting the parser with Megaparsec... For
now, we simply introduce a new language element: the `@` operator,
which is only used for struct initialization. This fixes the last
remaining shift/reduce conflict.