overhaul of the AST
This commit changes the approach wrt. Trees That Grow: rather than
having a bunch of extension types for each possible constructor, we
annotate the most important part of the tree: declarations,
statements, and expressions.

Furthermore, this commit expands the test suite and fixes a few
remaining errors, such as using `NonEmpty` for struct initialization.